python_2025/requirements.txt
Pillow>=9.0.0
tkinterdnd2>=0.3.0
pytest>=7.0.0
numpy>=1.20.0